pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Eduardo da Silva", email = "eduardo.silva@gmail.com"},
]
dependencies = [
    "django>=4.0.4",
    "djangorestframework>=3.13.1",
    "django-extensions>=3.1.5",
    "dependency-injector>=4.39.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=1.6.0",
    "pylint>=2.13.9",
    "pytest>=7.1.2",
    "flake8>=4.0.1",
    "black>=22.3.0",
    "pytest-cov>=3.0.0",
]
[tool.pdm.scripts]
manage = "python ./src/django_app/manage.py"
pep8 = "autopep8 --in-place --recursive ./src"
pylint = "pylint ./src"
runserver = "pdm run manage runserver 0.0.0.0:8000"
test = "pytest --ignore __pypackages__"
test_cov = "pdm run test --cov ./src --cov-fail-under 80"
test_cov_html = "pdm run test_cov --cov-report html:./__coverage"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
